// modules/transaction-manager/enums/service-code.enum.ts
export enum ETMServiceCode {
  OWN_TRANSFER = 'transfer',
  OWN_TRANSFER2 = 'transfer2',
  THIRD_TRANSFER = 'ttr-transf',
  THIRD_TRANSFER_MANAGER = 'ttr-admin',
  THIRD_TRANSFER_MANAGER2 = 'ttr-admin2',
  THIRD_TRANSFER2 = 'ttr-trans2',
  ACH_TRANSFER = 'ach-cnsadm',
  ACH_TRANSFER2 = 'ach-trans2',
  ACH_TRANSFER_MANAGER = 'ach-transf',
  DONATION = 'ope-dncs',
  DONATION2 = 'ope-dncs2',
  ACH_MASSIVE_TRANSFERENCE = 'ach-cararc',
  ACH_MASSIVE_TRANSFERENCE2 = 'web-cararc',
  THIRD_PARTY_PAYMENT_LOAN = 'ptr-pagpt2',
  ACH_MASSIVE_TRANSFERENCE_365 = 'ach-car365',
  PAYMENT_TAXES = 'PAYMENT_TAXES',
  PAYMENT_CHECKS = 'PAYMENT_CHECKS',
  PAYMENT_OF_PAYROLL = 'pla-aut',
  PAYMENT_OF_PAYROLL_SPECIAL = 'pla-wal',
  PAYMENT_OF_CREDIT_CARD = 'tc-pago',
  PAYMENT_OF_CREDIT_CARD_THIRD_PARTIES = 'pago-tc',
  PAYMENT_OF_SERVICES = 'mpg-cpago',
  PAYMENT_OF_TAXES = 'ope-pdet',
  PAYMENT_OF_PROVIDERS = 'pro-pag',
  BISV_PAYMENT_SUPPLIER = 'pag-pro',
  EXCHANGE_CURRENCY_AUCTION = 'ope-subdiv',
  EXCHANGE_CURRENCY_PURCHASE = 'ope-cmpdiv',
  CASHIER_CHECK_REQUEST = 'ges-solchq',
  DISBURSEMENT_OF_CREDIT = 'adq-desemb',
  DIVIDE_ALL = 'tc-divtodo',
  BULK_SERVICES_PAYMENT = 'mpg-cpmpg',
  LOAN_PAYMENT_PROFFER = 'ptr-grptmo',
  PAYMENT_APF_CONFIA_NPE = 'pag-afpnpe',
  PAYMENT_APF_CRECER_NPE = 'pag-crenpe',
  TREASURY_PAYMENT_NPE = 'ope-hacnpe',
  ACH_365_SIPA = 'web-sipa',
}

export const ETMServicesCode = {
  [ETMServiceCode.OWN_TRANSFER]: ETMServiceCode.OWN_TRANSFER,
  [ETMServiceCode.OWN_TRANSFER2]: ETMServiceCode.OWN_TRANSFER,
  [ETMServiceCode.DONATION2]: ETMServiceCode.DONATION,
  [ETMServiceCode.DONATION]: ETMServiceCode.DONATION,
  [ETMServiceCode.THIRD_TRANSFER]: ETMServiceCode.THIRD_TRANSFER,
  [ETMServiceCode.THIRD_TRANSFER_MANAGER]: ETMServiceCode.THIRD_TRANSFER,
  [ETMServiceCode.THIRD_TRANSFER_MANAGER2]: ETMServiceCode.THIRD_TRANSFER,
  [ETMServiceCode.THIRD_TRANSFER2]: ETMServiceCode.THIRD_TRANSFER,
}